Create README.md
This is an extremely simple local cafe website.
Simply follow the link to the list of cafes. You can even see a map!
Since this a list of cafes in Busan, South Korea, I removed the add cafe page as unneccesary, but I left both the code and index file intact.
I hope you enjoy it! Most of the cafes are near the beach, which is fantastic! The one in Seomeyon is beautiful and garden-like.